backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
fpdf2==2.7.8
qrcode[pil]==7.4.2
Pillow==10.2.0
aiosmtplib==3.0.1
email-validator==2.1.0
supabase==2.3.4
pydantic==2.5.3
python-multipart==0.0.6
jinja2==3.1.3
